Azure/requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
requests==2.31.0
